requirements.txt
# Civic Protocol Core - Python Dependencies

# Core API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic[email]>=2.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
sqlitedict>=2.1.0
alembic>=1.13.2

# Authentication
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
python-multipart>=0.0.6

# Core dependencies
requests>=2.28.0
cryptography>=3.4.8
pycryptodome>=3.15.0

# Web framework for dev node
flask>=2.2.0
flask-cors>=4.0.0

# Data handling
dataclasses-json>=0.5.7
python-dateutil>=2.8.2
PyYAML>=6.0.1
orjson>=3.10.0

# API utilities
slowapi>=0.1.9
python-dotenv>=1.0.1

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Optional: Advanced cryptography
# pyzkp>=0.1.0  # For zero-knowledge proofs
# semaphore>=0.1.0  # For rate limiting proofs













